Removed duplication of FormData interface by centralizing its usage (#33)
