+++ src/types/containers/guest-home-page/signup-dialog/SignUpDialog.types.ts
@@ -1,13 +1,5 @@
 import { UserRoleEnum } from '~/types'
 
-export interface FormData {
-  firstName: string
-  lastName: string
-  email: string
-  password: string
-  confirmPassword: string
-}
-
 export interface SignUpDialogProps {
   initialRole: UserRoleEnum
 }
